* sessionization logic 구현 - pyspark
* 스키마 정보 분리
